requirements.txt
tensorflow==2.13.0
streamlit==1.29.0
numpy==1.24.3
pillow
opencv-python